+++ cardano-cli/src/Cardano/CLI/Commands.hs
@@ -15,6 +15,7 @@ import           Cardano.CLI.Commands.Ping (PingCmd (..))
 import           Cardano.CLI.Compatible.Commands
 import           Cardano.CLI.EraBased.Commands
 import           Cardano.CLI.EraBased.Commands.Query
+import           Cardano.CLI.EraBased.Commands.TopLevelCommands
 import           Cardano.CLI.Legacy.Commands
 
 import           Options.Applicative.Types (ParserInfo (..), ParserPrefs (..))
